+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- [fixed] Fixing a regression introduced in 3.2.0, where `VerifyIDToken()`
+  cannot be used in App Engine.
+
 # v3.2.0
 
 - [added] Implemented the ability to create custom tokens without